Change Raid score int to long

+++ Lib9c/Action/Raid.cs
@@ -167,7 +167,7 @@ namespace Nekoyume.Action
             itemSlotState.UpdateCostumes(CostumeIds);
             states = states.SetState(itemSlotStateAddress, itemSlotState.Serialize());
 
-            int previousHighScore = raiderState.HighScore;
+            long previousHighScore = raiderState.HighScore;
             WorldBossState bossState;
             WorldBossGlobalHpSheet hpSheet = sheets.GetSheet<WorldBossGlobalHpSheet>();
             if (states.TryGetState(worldBossAddress, out List rawBossState))
@@ -252,7 +252,7 @@ namespace Nekoyume.Action
                 equipmentList, costumeList,
                 runeOptions, avatarState.level,
                 characterRow, costumeStatSheet);
-            int score = simulator.DamageDealt;
+            long score = simulator.DamageDealt;
             raiderState.Update(avatarState, cp, score, PayNcg, context.BlockIndex);
 
             // Reward.
